Refactor query builder into a base class
